rollback the total for backward compatible

+++ internal/api/shared/dto/token.go
@@ -53,7 +53,7 @@ type PaginatedOwners struct {
 type TokenListResponse struct {
 	Tokens []TokenResponse `json:"items"`
 	Offset *uint64         `json:"offset,omitempty"`
-	Total  int             `json:"total"`
+	Total  uint64          `json:"total"`
 }
 
 // MapTokenToDTO maps a schema.Token to TokenResponse
